refactor: add template to the naming conventions
