+++ src/app/[soorah]/SoorahAyah.template.tsx
@@ -1,37 +1,31 @@
 import Link from 'next/link'
 
 import { ExternalLink } from 'lucide-react'
-import Highlighter from 'react-highlight-words'
 
 import { Sajda } from '@/components/Sajda/Sajda'
 import { buildUrl } from '@/helpers/buildUrl'
 import { DisplayData } from '@/helpers/types'
 
 
-export type SearchAyahProps = {
+export type SoorahAyahProps = {
   data: DisplayData
   sajda?: number[]
-  mark?: string
 }
 
-export const SearchAyah = ({ data, sajda, mark = '' }: SearchAyahProps) => (
+export const SoorahAyahTemplate = ({ data, sajda }: SoorahAyahProps) => (
   <li className="soorah-list-item">
     <div className="flex flex-row">
       <span className="badge">
-        {data.soorah}:{data.ayah}
+        {data.ayah}
         {sajda?.includes(data.ayah) && <Sajda />}
-      </span>{' '}
-      <Highlighter
-        searchWords={[mark]}
-        textToHighlight={data.content}
-        autoEscape={true}
-        highlightClassName="bg-warning"
-      />
+      </span>
+      <span>{data.content}</span>
     </div>
     <Link
       href={buildUrl(data.soorah, data.ayah, data.translator)}
       className="read-ayah"
       prefetch={false}
+      title="oxu"
     >
       <ExternalLink />
     </Link>
