+++ src/app/search/[search]/SearchAyah.template.tsx
@@ -1,31 +1,37 @@
 import Link from 'next/link'
 
 import { ExternalLink } from 'lucide-react'
+import Highlighter from 'react-highlight-words'
 
 import { Sajda } from '@/components/Sajda/Sajda'
 import { buildUrl } from '@/helpers/buildUrl'
 import { DisplayData } from '@/helpers/types'
 
 
-export type SoorahAyahProps = {
+export type SearchAyahProps = {
   data: DisplayData
   sajda?: number[]
+  mark?: string
 }
 
-export const SoorahAyah = ({ data, sajda }: SoorahAyahProps) => (
+export const SearchAyahTemplate = ({ data, sajda, mark = '' }: SearchAyahProps) => (
   <li className="soorah-list-item">
     <div className="flex flex-row">
       <span className="badge">
-        {data.ayah}
+        {data.soorah}:{data.ayah}
         {sajda?.includes(data.ayah) && <Sajda />}
-      </span>
-      <span>{data.content}</span>
+      </span>{' '}
+      <Highlighter
+        searchWords={[mark]}
+        textToHighlight={data.content}
+        autoEscape={true}
+        highlightClassName="bg-warning"
+      />
     </div>
     <Link
       href={buildUrl(data.soorah, data.ayah, data.translator)}
       className="read-ayah"
       prefetch={false}
-      title="oxu"
     >
       <ExternalLink />
     </Link>
